refactor(resolve): fold ThroughRowProjection into the engine tier
The generic, shape-agnostic ThroughRowProjection is the primary implementation
of the Projection Protocol, so it belongs beside reconcile/diff/Projection in
the catalog-free _engine.py, not in the domain module _relationships.py. Moving
it (with its column converters) keeps the whole movable mechanism tier in one
import-linter-pinned module, so POST4's hoist to provenance is a one-file
git mv. The concrete builders and the bespoke Alias/Media projections name
domain models and stay put. Pure relocation; no behavior change.

Also update the plan's POST4 note to record the now-settled home.
